requirements.txt
absl-py==1.0.0
appdirs==1.4.4
audioread==2.1.9
blis==0.7.8
cachetools==5.0.0
captum==0.6.0
catalogue==2.0.8
charset-normalizer==2.0.12
click==8.1.3
clip
cloudpickle==2.0.0
cycler==0.11.0
cymem==2.0.6
decord==0.6.0
ffmpeg==1.4
ffmpeg-python==0.2.0
fonttools==4.29.1
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5.post20220506
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
importlib-metadata==4.11.1
iniconfig==1.1.1
iopath==0.1.9
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
langdetect==1.0.9
librosa==0.9.1
llvmlite==0.38.0
markdown==3.3.6
matplotlib==3.5.1
murmurhash==1.0.7
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
pandas==1.4.1
pathy==0.6.2
pillow==9.0.1
pluggy==1.0.0
pooch==1.6.0
portalocker==2.4.0
preshed==3.0.6
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pytest==7.1.2
pytz==2021.3
pyyaml==6.0
regex==2022.9.13
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
smart-open==5.2.1
soundfile==0.10.3.post1
spacy==3.4.1
spacy-language-detection==0.2.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
srsly==2.4.4
submitit==1.4.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
thinc==8.1.0
threadpoolctl==3.1.0
timm==0.3.2
tomli==2.0.1
torch-tb-profiler==0.3.1
torchaudio==0.12.1
tqdm==4.62.3
typer==0.4.2
urllib3==1.26.8
wasabi==0.10.1
werkzeug==2.0.3
wget==3.2
yacs==0.1.8
